backend/requirements.txt
Flask
flask
Flask-SQLAlchemy
Flask-WTF
requests
mailjet-rest
flask-migrate
pymysql
flask-login
flask-bcrypt
marshmallow
python-dotenv
Flask-Cors